Vampiric effect improvement. Contributed by Sdw-.

+++ L2J_Mobius_Underground/java/com/l2jmobius/gameserver/model/actor/stat/CharStat.java
@@ -63,7 +63,7 @@ public class CharStat
 	private final int[] _traitsInvul = new int[TraitType.values().length];
 	/** Creature's maximum buff count. */
 	private int _maxBuffCount = Config.BUFFS_MAX_AMOUNT;
-	private int _vampiricSum = 0;
+	private double _vampiricSum = 0;
 	
 	private final Map<Stats, Double> _statsAdd = new EnumMap<>(Stats.class);
 	private final Map<Stats, Double> _statsMul = new EnumMap<>(Stats.class);
@@ -870,12 +870,12 @@ public class CharStat
 		});
 	}
 	
-	public void addToVampiricSum(int sum)
+	public void addToVampiricSum(double sum)
 	{
 		_vampiricSum += sum;
 	}
 	
-	public int getVampiricSum()
+	public double getVampiricSum()
 	{
 		_lock.readLock().lock();
 		try
